backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-docx==1.1.0
PyMuPDF==1.23.8
gunicorn==21.2.0
